myproj/requirements.txt
Flask==1.1.2
Flask-WTF==0.14.3
mysql-connector-python==8.0.21
gunicorn==20.0.4
redis==3.5.3
